backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for NanoBanana API
httpx==0.25.2

# File handling and image processing
python-multipart==0.0.6
Pillow==10.1.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # for testing

# Code quality
ruff==0.1.6
mypy==1.7.1

# Environment variables
python-dotenv==1.0.0
